# backend/docs/swagger.yaml
basePath: /api
definitions:
  backend-api_internal_models.Jurusan:
    properties:
      fakultas:
        type: string
      id_jurusan:
        type: integer
      jenjang:
        type: string
      nama_jurusan:
        type: string
    type: object
  backend-api_internal_models.Mahasiswa:
    properties:
      alamat:
        type: string
      id:
        type: integer
      id_jurusan:
        type: integer
      jurusan:
        $ref: '#/definitions/backend-api_internal_models.Jurusan'
      nama:
        type: string
      nim:
        type: string
      tgl_lahir:
        type: string
      umur:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Ini adalah dokumentasi REST API untuk mengelola data mahasiswa dan
    jurusan.
  title: API Manajemen Mahasiswa
  version: "1.0"
paths:
  /jurusan:
    get:
      consumes:
      - application/json
      description: Mengambil seluruh data referensi jurusan untuk dropdown
      produces:
      - application/json
      responses:
        "200":
          description: Data jurusan berhasil diambil
          schema:
            items:
              $ref: '#/definitions/backend-api_internal_models.Jurusan'
            type: array
        "500":
          description: Gagal mengambil data jurusan
          schema:
            additionalProperties:
              type: string
            type: object
      summary: GetJurusan
      tags:
      - Jurusan
  /mahasiswa:
    get:
      description: Mengambil semua data mahasiswa beserta detail jurusannya
      produces:
      - application/json
      responses:
        "200":
          description: Daftar data mahasiswa
          schema:
            items:
              $ref: '#/definitions/backend-api_internal_models.Mahasiswa'
            type: array
      summary: Mengambil Data Mahasiswa
      tags:
      - mahasiswa
    post:
      consumes:
      - application/json
      description: Menambahkan data mahasiswa baru ke database
      parameters:
      - description: Data Mahasiswa JSON
        in: body
        name: mahasiswa
        required: true
        schema:
          $ref: '#/definitions/backend-api_internal_models.Mahasiswa'
      produces:
      - application/json
      responses:
        "201":
          description: Data mahasiswa berhasil disimpan
          schema:
            $ref: '#/definitions/backend-api_internal_models.Mahasiswa'
        "400":
          description: Format data tidak valid
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Menambahkan Data Mahasiswa
      tags:
      - mahasiswa
  /mahasiswa/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus data mahasiswa berdasarkan ID
      parameters:
      - description: ID Mahasiswa
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Data mahasiswa berhasil dihapus
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Mahasiswa dengan ID tersebut tidak ditemukan
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Menghapus Data Mahasiswa
      tags:
      - mahasiswa
    put:
      consumes:
      - application/json
      description: Mengupdate data mahasiswa berdasarkan ID
      parameters:
      - description: ID Mahasiswa
        in: path
        name: id
        required: true
        type: integer
      - description: Data Mahasiswa JSON
        in: body
        name: mahasiswa
        required: true
        schema:
          $ref: '#/definitions/backend-api_internal_models.Mahasiswa'
      produces:
      - application/json
      responses:
        "200":
          description: Data mahasiswa berhasil diperbarui
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Mahasiswa dengan ID tersebut tidak ditemukan
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Mengupdate Data Mahasiswa
      tags:
      - mahasiswa
  /mahasiswa/export:
    get:
      description: Mengunduh seluruh daftar data mahasiswa ke dalam file Excel (.xlsx)
        dengan format tanggal bersih (YYYY-MM-DD).
      produces:
      - application/octet-stream
      responses:
        "200":
          description: File Laporan_Mahasiswa.xlsx berhasil diunduh
          schema:
            type: file
        "500":
          description: Gagal memproses data atau menulis file Excel
          schema:
            additionalProperties: true
            type: object
      summary: Export Data Mahasiswa ke Excel
      tags:
      - mahasiswa
schemes:
- http
swagger: "2.0"